requirements.txt
psycopg2
sqlalchemy
pandas
matplotlib
elasticsearch
sentence_transformers
sklearn.cluster